Integrated the Alpha Vantange DataFrame extension.
The timeserise data frome Alpha Vantage does not have the adjusted closing price column because that is not part of the FREE API.  The cheapestmonth price for their premium service is $25/mo.  That's too expensive for casual users.  So I'm faking an adjusted closing price by just duplicating the closing price.   It should be good enought for work with recent data.