Did some house cleaning, reorganization.  Still do not know why the hashie extensions dash PredefinedValues constant is undefined.  Commented out the include in the Config class.  It was only being used for the log_level property.  I should be able to work around that.
Will be integrating the "alphavantage" gem to replace those places where I am currently calling the Alpha Vantage API directory... that's mostly in the Stock class.  The reason I will be using the gem, is because I intend to update that gem with the shopify limiter to support rate limiting access to the AV apis when using a free api key.

I also intend to create a new gem to front the dumbstockapi website at dumbstockapi.com
Current I'm only using it in the Ticker class to download ticker symbols to a CSV file then using that CSV dump as a Hash to validate ticker values in the Stock class.